refactor: docker 파일 수정 및 멤버 디렉토리 구조 변경 - #21
