+++ Dockerfile
@@ -1,11 +1,5 @@
-# 기본 Java 이미지 사용
+# doker 이미지 생성을 위한 스크립트
+
 FROM openjdk:17
-
-# JAR 파일 경로를 인자로 받음
-ARG JAR_FILE=build/libs/app.jar
-
-# 파일을 컨테이너의 app.jar로 복사
-COPY ${JAR_FILE} /app.jar
-
-# 컨테이너 시작 시 Java 애플리케이션 실행
+COPY build/libs/fmhj-0.0.1-SNAPSHOT.jar /app.jar
 ENTRYPOINT ["java", "-jar", "/app.jar"]
